requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
openpyxl>=3.1.0
google-genai>=1.0.0
